+++ CHANGELOG.md
@@ -6,6 +6,8 @@ In Development
 - Fix a stall when cleaning up after an error
 - Add `--trace-progress` option
 - Drastically reduce memory usage
+- Reject keys with directory components with special meaning, not just keys
+  with filenames with special meaning
 
 v0.1.0-alpha (2024-11-26)
 -------------------------